fix state management and switch to redux
