+++ src/store/ac.js
@@ -1,0 +1,3 @@
+import {Constance} from './constance';
+
+export const deleteClickHandler = (id) => ({type: Constance.DELETE_CLICK_HANDLER, payload: id});
