+++ src/store/constance.js
@@ -1,0 +1,3 @@
+export const Constance = {
+    DELETE_CLICK_HANDLER: 'DELETE_CLICK_HANDLER',
+}
