+++ src/store/redux-store.js
@@ -1,0 +1,8 @@
+import {combineReducers, createStore} from "redux";
+import {reducer} from "./reducers";
+
+
+const reducers = combineReducers({
+    data: reducer
+})
+export const store = createStore(reducers);
